+++ src/bbf/runner.py
@@ -58,7 +58,6 @@ def parse(tokens: list[Token]) -> ProgTopLevelStmt:
 def generate_exe(
     prog: ProgTopLevelStmt, exe_path: Path, *, verbose: bool = False
 ) -> None:
-    assert exe_path.is_file()
     main_emitter = Emitter()
     nasm_macros_emitter = Emitter()
     asm_codegen = AsmCodeGen(main_emitter, nasm_macros_emitter)